Moved player searching UI to a separate component; minor corrections to the backend
